Create Day-5.c++
Day 5 of 30 🗓️

Problem : Find the Leaders in an Array
given an integer array arr of size n. An element is considered a leader if it is greater than all the elements to its right. Task is to find all such leader elements in the array.

Input :
An integer array arr of size n.
Example : 
arr = [16, 17, 4, 3, 5, 2]

Output :
Return an array containing all the leader elements in the order in which they appear in the original array.
Example:
Leaders: [17, 5, 2]